backend/requirements.txt
espn-api>=0.33.0
pandas>=2.0.0
openpyxl>=3.1.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
